requirements.txt
langchain==0.1.12
langchain-core==0.1.32
langchain-google-genai==0.0.9
langgraph==0.0.29
chromadb==0.4.22
python-dotenv==1.0.1
requests==2.31.0
google-generativeai==0.3.2
streamlit==1.31.1
